+++ .idea/UDP-TCP-Chat.iml
@@ -3,7 +3,7 @@
   <component name="NewModuleRootManager" inherit-compiler-output="true">
     <exclude-output />
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/netcat-release/Java - netcat TCP/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/Exercise 4" isTestSource="false" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
