+++ Exercise 4/TCP_Chat_Client.java
@@ -1,0 +1,116 @@
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.PrintWriter;
+import java.net.Socket;
+import java.net.UnknownHostException;
+import java.util.HashMap;
+import java.util.Map;
+
+public class TCP_Chat_Client {
+    private static String name;
+    private static String serverIP;
+    private static int serverPort;
+
+    private static HashMap<String, Integer> functions = new HashMap<>(){{
+        put("send",3);
+        put("getClientList",1);
+        put("ask",3);
+    }};
+
+    private static Map<String, String> questions = new HashMap<>() {{
+        put("What is the capital of Germany?", "Berlin");
+        put("What is the capital of France?", "Paris");
+        put("Bestes Bier?", "Bitburger");
+    }};
+
+    private static void fatal(String input) {
+        System.err.println(input);
+        System.exit(-1);
+    }
+
+    public static boolean isIP(String ip) { // Checks if String is valid IPv4 address
+        String[] parts = ip.split("\\."); // Split by dot
+        if (parts.length != 4) { return false; } // Must be 4 chunks
+        for (String p : parts) { // Check if numbers are valid
+            try {
+                int number = Integer.parseInt(p);
+                if (number < 0 || number > 255) { return false; }
+            } catch (NumberFormatException e) { return false; }
+        }
+        return true;
+    }
+
+    public static boolean isPort(String port) {
+        try {
+            int number = Integer.parseInt(port);
+            if (number < 0 || number > 65535) { return false; }
+        } catch (NumberFormatException e) { return false; }
+        return true;
+    }
+
+    public static void main(String[] args) {
+
+        // Handling arguments, checking validity
+        if (args.length != 3) {
+            fatal("Arguments: \"<server ip address> <server port number> <client name>\"");
+        }
+        if (!isIP(args[0])) {
+            fatal("Invalid IP address");
+        } else {
+            serverIP = args[0];
+        }
+        if (!isPort(args[1])) {
+            fatal("Invalid port number");
+        } else {
+            serverPort = Integer.parseInt(args[1]);
+        }
+        name = args[2];
+
+        try (Socket socket = new Socket(serverIP, serverPort);
+             PrintWriter out = new PrintWriter(socket.getOutputStream(), true);
+             BufferedReader in = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+             BufferedReader stdIn = new BufferedReader(new InputStreamReader(System.in))) {
+            // closes automatically
+
+            new Thread(() -> {
+                try {
+                    String fromServer;
+                    while ((fromServer = in.readLine()) != null) {
+                        System.out.println(fromServer);
+                        String[] parts = fromServer.split(":",2);
+                        parts[1] = parts[1].trim();
+                        if (questions.containsKey(parts[1])) {
+                            String message = "send " + parts[0].split(" ",3)[2] + " " + questions.get(parts[1]);
+                            out.println(message);
+                            System.out.println("Answered question with " + questions.get(parts[1]));
+                        }
+                    }
+                } catch (IOException e) {
+                    fatal("Unable to get message from Server.");
+                }
+            }).start();
+
+            // Register the client with the server
+            out.println("register " + name);
+
+            System.out.println(name + " is connected to Server at IP " + serverIP + " on port " + serverPort + ".\nUse \"send <client name> <message>\" to send a message to a client.");
+
+            String userInput;
+            while ((userInput = stdIn.readLine()) != null) {
+                String[] parts = userInput.split(" ", 3);
+                if (functions.containsKey(parts[0]) && functions.get(parts[0]) == parts.length) {
+                    out.println(userInput);
+                    System.out.println("Message sent.");
+                } else {
+                    System.err.println("Unknown command. Possible commands:\n" + functions.keySet());
+                    }
+                }
+        } catch (UnknownHostException e) {
+            fatal("Unknown Server with IP " + serverIP);
+        } catch (IOException e) {
+            fatal("Unable to send message.");
+        }
+        System.exit(0);
+    }
+}
