--- netcat-release/Java - netcat UDP/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: oxoo2a.Main
-
